+++ src/Api/Config/ConfigManager.php
@@ -12,7 +12,7 @@
 namespace Puli\Manager\Api\Config;
 
 use Puli\Manager\Api\InvalidConfigException;
-use Puli\Manager\Api\Storage\StorageException;
+use Puli\Manager\Api\Storage\WriteException;
 use Webmozart\Expression\Expression;
 
 /**
@@ -41,7 +41,7 @@ interface ConfigManager
      *
      * @throws NoSuchConfigKeyException If the configuration key is invalid.
      * @throws InvalidConfigException   If the value is invalid.
-     * @throws StorageException         If the file cannot be written.
+     * @throws WriteException           If the file cannot be written.
      */
     public function setConfigKey($key, $value);
 
@@ -54,7 +54,7 @@ interface ConfigManager
      *
      * @throws NoSuchConfigKeyException If a configuration key is invalid.
      * @throws InvalidConfigException   If a value is invalid.
-     * @throws StorageException         If the file cannot be written.
+     * @throws WriteException           If the file cannot be written.
      */
     public function setConfigKeys(array $values);
 
@@ -66,7 +66,7 @@ interface ConfigManager
      * @param string $key The removed configuration key.
      *
      * @throws NoSuchConfigKeyException If the configuration key is invalid.
-     * @throws StorageException         If the file cannot be written.
+     * @throws WriteException           If the file cannot be written.
      */
     public function removeConfigKey($key);
 
@@ -78,7 +78,7 @@ interface ConfigManager
      * @param Expression $expr The search criteria.
      *
      * @throws NoSuchConfigKeyException If a configuration key is invalid.
-     * @throws StorageException         If the file cannot be written.
+     * @throws WriteException           If the file cannot be written.
      */
     public function removeConfigKeys(Expression $expr);
 
@@ -87,7 +87,7 @@ interface ConfigManager
      *
      * The file is saved directly after removing the keys.
      *
-     * @throws StorageException If the file cannot be written.
+     * @throws WriteException If the file cannot be written.
      */
     public function clearConfigKeys();